+++ database/migrations/2017_09_01_000000_create_revenues_table.php
@@ -17,9 +17,9 @@ class CreateRevenuesTable extends Migration
             $table->integer('company_id');
             $table->integer('account_id');
             $table->date('paid_at');
-            $table->double('amount', 15, 4);
+            $table->decimal('amount', 15, 4);
             $table->string('currency_code');
-            $table->double('currency_rate', 15, 8);
+            $table->decimal('currency_rate', 15, 8);
             $table->integer('customer_id')->nullable();
             $table->text('description')->nullable();
             $table->integer('category_id');
