+++ database/migrations/2017_10_11_000000_create_invoice_totals_table.php
@@ -23,7 +23,7 @@ class CreateInvoiceTotalsTable extends Migration
             $table->integer('invoice_id');
             $table->string('code')->nullable();
             $table->string('name');
-            $table->double('amount', 15, 4);
+            $table->decimal('amount', 15, 4);
             $table->integer('sort_order');
             $table->timestamps();
             $table->softDeletes();
